% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breastcancer.R
\docType{data}
\name{breastcancer}
\alias{breastcancer}
\title{Breast Cancer Data}
\format{
\subsection{\code{breastcancer}}{

A tibble with 192 rows and 3 columns:
\describe{
\item{death}{Death, binary: 0, 1}
\item{stage}{Cancer stage, 3 categories}
\item{receptor}{Hormone receptor status, binary: "High", "Low"}
...
}
}
}
\source{
Newman SC. Biostatistical methods in epidemiology.
New York, NY: Wiley, 2001, table 5.3
}
\usage{
breastcancer
}
\description{
A cohort of women with breast cancer and complete follow-up, as used by
Spiegelman and Hertzmark (Am J Epidemiol 2005) and
Greenland (Am J Epidemiol 2004).
}
\keyword{datasets}
