% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margstd-utils.R
\name{confint.margstd}
\alias{confint.margstd}
\title{Bias-corrected/accelerated bootstrap confidence intervals}
\usage{
\method{confint}{margstd}(object, parm = NULL, level = 0.95, bootrepeats = 200, ...)
}
\arguments{
\item{object}{Model fitted through marginal standardization}

\item{parm}{Not used, for compatibility}

\item{level}{Confidence level, defaults to 0.95.}

\item{bootrepeats}{Bootstrap repeats. Defaults to 200. Strongly recommend >1000.}

\item{...}{Not used}
}
\value{
Matrix: First column, lower bound; second column, upper bound.
}
\description{
Confidence intervals for models fit using marginal standardization.
}
