% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margstd_delta.R
\name{confint.margstd_delta}
\alias{confint.margstd_delta}
\title{Delta method confidence intervals}
\usage{
\method{confint}{margstd_delta}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{Model fitted through marginal standardization, delta method}

\item{parm}{Not used, for compatibility}

\item{level}{Confidence level, defaults to \code{0.95}.}

\item{...}{Not used}
}
\value{
Matrix: First column, lower bound; second column, upper bound.
}
\description{
Confidence intervals for models fit using marginal standardization
based on the melta method.
}
