% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robpoisson-utils.R
\name{confint.robpoisson}
\alias{confint.robpoisson}
\title{Robust confidence intervals for Poisson model}
\usage{
\method{confint}{robpoisson}(object, parm = NULL, level = 0.95, ...)
}
\arguments{
\item{object}{Fitted model}

\item{parm}{Not used}

\item{level}{Confidence level, defaults to 0.95}

\item{...}{Additional arguments, not used}
}
\value{
Matrix: First column, lower bound; second column, upper bound.
}
\description{
Estimate confidence intervals for the Poisson model
with robust/sandwich/empirical covariance structure.
}
