% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm-utils.R
\name{conv.test}
\alias{conv.test}
\title{Helper function for logbin and addreg packages}
\usage{
conv.test(theta1, theta2, epsilon)
}
\arguments{
\item{theta1}{theta1}

\item{theta2}{theta2}

\item{epsilon}{epsilon}
}
\value{
Logical.
}
\description{
This is a 1:1 copy of a function from logbin that is needed but not exported
by logbin and addreg.
}
\keyword{internal}
