% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimation.R
\name{print.risks}
\alias{print.risks}
\title{Print model}
\usage{
\method{print}{risks}(x, ...)
}
\arguments{
\item{x}{Fitted model}

\item{...}{Passed to print.glm()}
}
\value{
No return value, called for printing
}
\description{
Print fitted risks model. The only change, compared to \code{print.glm()},
is the addition of the main type of model: relative risk or risk difference.
If multiple models were fitted via \code{approach = "all"}, then the
first converged model will be printed.
}
