% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimation.R
\name{print.summary.risks}
\alias{print.summary.risks}
\title{Print model summary}
\usage{
\method{print}{summary.risks}(x, conf.int = TRUE, default = TRUE, ...)
}
\arguments{
\item{x}{Model}

\item{conf.int}{Add confidence intervals to printout? Defaults to TRUE.}

\item{default}{Normal confidence intervals via confint.default()? Default to TRUE.
By setting \code{default = FALSE}, profile likelihood-based
confidence intervals can be calculated for binomial models.}

\item{...}{Passed on}
}
\description{
Print summaries for "risks" models. The printout is the same as
for regular summaries of generalized linear models fit via
\code{stats::glm()}, except that the type of "risks" model
is printed first (e.g., "Poisson model with robust covariance")
and confidence intervals for model parameters are printed at the end.
}
