% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_risk.R
\name{riskratio}
\alias{riskratio}
\alias{riskdiff}
\title{Fit risk ratio and risk difference models}
\usage{
riskratio(
  formula,
  data,
  approach = c("auto", "all", "robpoisson", "glm", "glm_start", "glm_cem",
    "glm_cem_start", "margstd", "logistic"),
  variable = NULL,
  at = NULL,
  ...
)

riskdiff(
  formula,
  data,
  approach = c("auto", "all", "robpoisson", "glm", "glm_start", "glm_cem",
    "glm_cem_start", "margstd", "logistic"),
  variable = NULL,
  at = NULL,
  ...
)
}
\arguments{
\item{formula}{A formula object of the form 'response ~ predictors'.}

\item{data}{A \code{tibble} or \code{data.frame} object.}

\item{approach}{Optional: Method for model fitting.
\itemize{
\item \code{"auto"} (default) is recommended; it selects the most efficient
approach that converges and ensures that predicted probabilities are
within range (< 1; see Details).
\item \code{"all"} will attempt to fit
the model via all implemented approaches to allow for comparisons.
}

The other options allow for directly selecting a fitting approach,
some of which may not converge or yield out-of-range predicted probabilities.
See full documentation (currently at \url{https://github.com/stopsack/risks}) for details.
\itemize{
\item \code{"glm"} Binomial model.
\item \code{"glm_start"} Binomial model with starting values from Poisson model.
\item \code{"robpoisson"} Poisson model with robust covariance.
\item \code{"glm_cem"} Binomial model fitted with combinatorial expectation maximization.
\item \code{"glm_cem_start"} As \code{glm_cem}, with Poisson starting values.
\item \code{"margstd"} Marginal standardization after logistic model.
}}

\item{variable}{Optional: variable to use for marginal standardization.
If \code{variable} is not provided and marginal standardization is
attempted, then the first binary or categorical variable in the model
is used as the exposure. Levels are determined automatically for
variables types \code{logical}, \code{character}, \code{factor},
and \code{numeric} (only if no more than 2 levels for the latter);
otherwise levels must be supplied via \code{at =}.}

\item{at}{Optional: Levels of variable \code{variable} for marginal
standardization. \code{at =} determines the levels at which contrasts of the exposure
are to be assessed. The level listed first is used as the reference.
Levels must exist in the data for character, factor or ordered factor variables.
For numeric variables, levels that do not exist in the data
can be interpolations or extrapolations; if levels exceed the
extremes of the data (extrapolation), a warning will be displayed.}

\item{...}{Optional: Further arguments passed to fitting functions (\code{glm},
\code{logbin}, or \code{addreg}).}
}
\value{
Fitted model. Pass this object to \code{\link{summary.risks}} to obtain
an overview of results; to \code{\link{tidy.risks}} to obtain a tibble of
coefficients and confidence intervals; or to \code{\link{predict.glm}(type = "response")}
to obtain fitted values (predictions).
}
\description{
\code{riskratio} and \code{riskdiff} provide a flexible interface to fitting
risk ratio and risk difference models.

In cohort studies with a binary outcome, risk ratios and risk differences
are typically more appropriate to report than odds ratios from logistic regression,
yet such models have historically been difficult to implement in standard software.

The risks package selects an efficient way to fit risk ratio or
risk difference models successfully, which will converge whenever logistic models converge.
Optionally, a specific approach to model fitting can also be requested.
Implemented are Poisson models with robust covariance, binomial models,
binomial models aided in convergence by starting values obtained
through Poisson models, binomial models fitted via
combinatorial expectation maximization (optionally also with Poisson starting values),
and estimates obtained via marginal standardization after logistic regression.

Adjusting for covariates (e.g., confounders) the model specification (\code{formula =}) is possible.
}
\section{Functions}{
\itemize{
\item \code{riskratio}: Fit risk ratio models

\item \code{riskdiff}: Fit risk difference models
}}

\examples{
# Newman SC. Biostatistical methods in epidemiology. New York, NY: Wiley, 2001, table 5.3
library(tibble)  # used to set up example data
dat <- tibble(
  death    = c(rep(1, 54), rep(0, 138)),
  stage    = c(rep("Stage I", 7),  rep("Stage II", 26), rep("Stage III", 21),
               rep("Stage I", 60), rep("Stage II", 70), rep("Stage III", 8)),
  receptor = c(rep("Low", 2),  rep("High", 5),  rep("Low", 9),  rep("High", 17),
               rep("Low", 12), rep("High", 9),  rep("Low", 10), rep("High", 50),
               rep("Low", 13), rep("High", 57), rep("Low", 2),  rep("High", 6)))

# Risk ratio model
fit_rr <- riskratio(formula = death ~ stage + receptor, data = dat)
fit_rr
summary(fit_rr)

# Risk difference model
fit_rd <- riskdiff(formula = death ~ stage + receptor, data = dat)
fit_rd
summary(fit_rd)
}
\references{
Wacholder S. Binomial regression in GLIM: Estimating risk ratios
and risk differences. Am J Epidemiol 1986;123:174-184.
(Binomial regression models; approach = "glm")

Spiegelman D, Hertzmark E. Easy SAS Calculations for Risk or
Prevalence Ratios and Differences. Am J Epidemiol 2005;162:199-200.
(Binomial models fitted used starting values from Poisson models;
approach = "glm_start")

Zou G. A modified Poisson regression approach to prospective
studies with binary data. Am J Epidemiol 2004;159:702-706.
(Poisson model with robust/sandwich standard errors;
approach = "robpoisson")

Donoghoe MW, Marschner IC. logbin: An R Package for
Relative Risk Regression Using the Log-Binomial Model.
J Stat Softw 2018;86(9). (Log-binomial models fitted via combinatorial
expectation maximization; estimate = "rr", approach = "glm_cem")

Donoghoe MW, Marschner IC. Stable computational methods
for additive binomial models with application to adjusted risk differences.
Comput Stat Data Anal 2014;80:184-96. (Additive binomial models
fitted via combinatorial expectation maximization;
estimate = "rd", approach = "glm_cem")

Localio AR, Margolis DJ, Berlin JA.
Relative risks and confidence intervals were easily computed
indirectly from multivariable logistic regression.
J Clin Epidemiol 2007;60(9):874-82. (Marginal standardization after fitting a
logistic model; approach = "margstd")
}
