% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risks.R
\docType{package}
\name{risks}
\alias{risks}
\title{risks: Estimate risk ratios and risk differences using regression}
\description{
The risks package allows for fitting risk ratio and risk difference
models using regression.
}
\section{risks functions}{

estimate_risk: main model fitting function
summary: summarize fitted model
tidy: return tibble (data frame) of parameters, coefficients, standard errors, confidence intervals
confint: return confidence intervals
}

