% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caseduplicate.R
\name{summary.duplicate}
\alias{summary.duplicate}
\title{Summary for logistic model with case duplication and cluster-robust covariance}
\usage{
\method{summary}{duplicate}(
  object,
  dispersion = NULL,
  correlation = FALSE,
  symbolic.cor = FALSE,
  ...
)
}
\arguments{
\item{object}{Model}

\item{dispersion}{Not used}

\item{correlation}{Not used}

\item{symbolic.cor}{Not used}

\item{...}{Other arguments, not used}
}
\value{
Model summary (list)
}
\description{
Summarize results from fitting a logistic model with case duplication and
cluster-robust covariance.
The output is the same as for a regular \code{summary(glm(...))},
except for using cluster-robust standard errors.
}
