% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margstd-utils.R
\name{summary.margstd}
\alias{summary.margstd}
\title{Summary for models using marginal standardization}
\usage{
\method{summary}{margstd}(
  object,
  dispersion = NULL,
  correlation = FALSE,
  symbolic.cor = FALSE,
  level = 0.95,
  bootrepeats = 200,
  ...
)
}
\arguments{
\item{object}{Model}

\item{dispersion}{Not used}

\item{correlation}{Not used}

\item{symbolic.cor}{Not used}

\item{level}{Confidence level, defaults to 95\%}

\item{bootrepeats}{Bootstrap repeats for standard errors. Defaults to 200. Strongly recommend >1000.}

\item{...}{Not used}
}
\value{
Model summary (list)


}
\description{
Summary for models using marginal standardization
}
