% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margstd_delta.R
\name{summary.margstd_delta}
\alias{summary.margstd_delta}
\title{Summary for models using marginal standardization with delta method SEs}
\usage{
\method{summary}{margstd_delta}(
  object,
  dispersion = NULL,
  correlation = FALSE,
  symbolic.cor = FALSE,
  level = 0.95,
  ...
)
}
\arguments{
\item{object}{Model}

\item{dispersion}{Not used}

\item{correlation}{Not used}

\item{symbolic.cor}{Not used}

\item{level}{Confidence level, defaults to \code{0.95}.}

\item{...}{Not used}
}
\value{
Model summary (list)
}
\description{
Summary for models using marginal standardization with delta method SEs
}
