% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robpoisson-utils.R
\name{summary.robpoisson}
\alias{summary.robpoisson}
\title{Summary for Poisson model with robust covariance}
\usage{
\method{summary}{robpoisson}(
  object,
  dispersion = NULL,
  correlation = FALSE,
  symbolic.cor = FALSE,
  ...
)
}
\arguments{
\item{object}{Model}

\item{dispersion}{Not used}

\item{correlation}{Not used}

\item{symbolic.cor}{Not used}

\item{...}{Other arguments, not used}
}
\value{
Model summary (list)
}
\description{
Summarize results from fitting a Poisson model with
robust/empirical/sandwich covariance.
The output is the same as for a regular \code{summary(glm(...))},
except for using robust standard errors.
}
