% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimation.R
\name{tidy.risks}
\alias{tidy.risks}
\title{Tidy model summaries for risks models}
\usage{
\method{tidy}{risks}(
  x,
  conf.int = TRUE,
  conf.level = 0.95,
  bootrepeats = 200,
  exponentiate = FALSE,
  default = TRUE,
  ...
)
}
\arguments{
\item{x}{Model}

\item{conf.int}{Show confidence intervals?}

\item{conf.level}{Confidence level}

\item{bootrepeats}{Bootstrap repeats. For marginal standardized models. >1000 recommended.}

\item{exponentiate}{Exponentiate? For log links}

\item{default}{Use normality-based confidence intervals?}

\item{...}{Passed on}
}
\value{
tibble
}
\description{
Obtain a tibble (data frame) with parameters,
coefficients, standard errors, confidence limits,
and p-values. Usual disclaimers about the use of the latter.
A column with the type of model fitted is also being added.
}
\details{
If running \code{tidy(...)} on a model using marginal
standardization via \code{approach = "margstd"},
by default, the number of bootstrap repeats is very low.
It is strongly recommended to increase them to >1000. See examples.

If multiple types of models are fitted, \code{tidy()} can be used
to parameters for all models at once, in one tibble. The last
column of the tibble includes the name of the model. See examples.
}
\examples{
# Define example data
library(tibble)  # for tibble() function
library(broom)   # required for tidy.risks()
dat <- tibble(
  id = 1:192,
  death = c(rep(1, 54), rep(0, 138)),
  stage = c(rep("Stage I", 7),  rep("Stage II", 26), rep("Stage III", 21),
            rep("Stage I", 60), rep("Stage II", 70), rep("Stage III", 8)),
  receptor = c(rep("Low", 2),  rep("High", 5),  rep("Low", 9),  rep("High", 17),
               rep("Low", 12), rep("High", 9),  rep("Low", 10), rep("High", 50),
               rep("Low", 13), rep("High", 57), rep("Low", 2),  rep("High", 6)))
# Fit and tidy the model
fit_rr <- estimate_risk(formula = death ~ stage + receptor, data = dat)
tidy(fit_rr)

# Marginal standardization,
# increase number of bootstrap repeats:
fit_rr <- estimate_risk(formula = death ~ stage + receptor, data = dat, approach = "margstd")
tidy(fit_rr, bootrepeats = 250)

# Multiple types of models fitted:
fit_rr <- estimate_risk(formula = death ~ stage + receptor, data = dat, approach = "all")
tidy(fit_rr)
}
