% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_risk.R
\name{estimate_risk}
\alias{estimate_risk}
\title{Fit risk ratio and risk difference models}
\usage{
estimate_risk(
  data,
  formula,
  estimate = "rr",
  level = 0.95,
  pvalue = FALSE,
  sourcecolumn = FALSE,
  verbose = TRUE,
  logistic = FALSE,
  addbin = TRUE,
  ...
)
}
\arguments{
\item{data}{a \code{tibble}(?) or \code{data.frame} object}

\item{formula}{a formula object}

\item{estimate}{the type of estimate to report: risk (prevalance) ratio
("rr") or risk difference ("rd")}

\item{level}{confidence interval width}

\item{pvalue}{a logical value indicating whether to add p-values to the table
of results}

\item{sourcecolumn}{a logical value indicating whether to add column with
source of final estimates (binomial vs. Poisson)}

\item{verbose}{a logical indicating whether to print all model summaries or
just final coefficients}

\item{logistic}{a logical indicating whether to show logistic regression for
comparison}

\item{addbin}{a logical indicating whether to attempt \code{addbin()} if
\code{glm()} fits fail to converge}

\item{...}{further arguments passed to \code{glm}}
}
\description{
\code{estimate_risk} Fits risk ratio and risk difference models using robust
Poisson and binomial models.
}
\details{
This is a paragraph which includes important methods details to be described.
}
\references{
Spiegelman D, Hertzmark E. Easy SAS Calculations for Risk or
Prevalence Ratios and Differences. Am J Epidemiol 2005;162:199-200.
(overview)

Wacholder S. Binomial regression in GLIM: Estimating risk ratios
and risk differences. Am J Epidemiol 1986;123:174-184. (log-binomial model)

Zou G. A modified Poisson regression approach to prospective
studies with binary data. Am J Epidemiol 2004;159:702-706. (Poisson with
sandwich SE)

Marschner's addbin package
}
