% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimation.R
\name{print.summary.risks}
\alias{print.summary.risks}
\title{Print model summary}
\usage{
\method{print}{summary.risks}(x, ...)
}
\arguments{
\item{x}{Model}

\item{...}{Passed on}
}
\value{
No return value, called for printing
}
\description{
Print summaries for "risks" models. The printout is the same as
for regular summaries of generalized linear models fit via
\code{stats::glm()}, except that the type of "risks" model
is printed first (e.g., "Poisson model with robust covariance")
and confidence intervals for model parameters are printed at the end.
}
