% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risks.R
\docType{package}
\name{risks}
\alias{risks}
\title{risks: Estimate risk ratios and risk differences using regression}
\description{
The risks package allows for fitting risk ratio and risk difference
models using regression.
}
\section{risks functions}{


\code{\link{riskratio}}: Fit risk ratio models.

\code{\link{riskdiff}}: Fit risk difference models.

\code{\link{summary.risks}}: Summarize fitted model.

\code{\link{tidy.risks}}: Tibble (data frame) of parameters, coefficients, standard errors, confidence intervals.

\code{\link{confint.robpoisson}}, \code{\link{confint.margstd}}: Confidence intervals. (Standard confidence intervals for generalized linear models are used for other models.)
}

\seealso{
\url{https://github.com/stopsack/risks}
}
