% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/risks.R
\docType{package}
\name{risks}
\alias{risks-package}
\alias{risks}
\title{risks: Estimate risk ratios and risk differences using regression}
\description{
The risks package allows for fitting risk ratio and risk difference
models using regression.
}
\section{risks functions}{


\code{\link[risks]{riskratio}}: Fit risk ratio models.

\code{\link[risks]{riskdiff}}: Fit risk difference models.

\code{\link[risks]{summary.risks}}: Summarize fitted model.

\code{\link[risks]{tidy.risks}}: Tibble (data frame) of parameters,
coefficients, standard errors, confidence intervals.

\code{\link[risks]{confint.risks}}: Confidence intervals.
}

\seealso{
\url{https://github.com/stopsack/risks}
}
\author{
\strong{Maintainer}: Konrad Stopsack \email{stopsack@post.harvard.edu} (\href{https://orcid.org/0000-0002-0722-1311}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Travis Gerke \email{tgerke@mail.harvard.edu} (\href{https://orcid.org/0000-0002-9500-8907}{ORCID})
}

}
