% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimation.R
\name{summary.risks}
\alias{summary.risks}
\title{Generate model summary}
\usage{
\method{summary}{risks}(object, ...)
}
\arguments{
\item{object}{Fitted model}

\item{...}{Passed on}
}
\description{
Determine type of "risks" model fitted and
generate appropriate summary.
}
