% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimation.R
\name{summary.risks}
\alias{summary.risks}
\title{Generate model summary}
\usage{
\method{summary}{risks}(object, conf.int = TRUE, default = TRUE, ...)
}
\arguments{
\item{object}{Fitted model}

\item{conf.int}{Add confidence intervals to printout? Defaults to TRUE.}

\item{default}{Normal confidence intervals via confint.default()?
Default to TRUE. By setting \code{default = FALSE}, profiling-based
confidence intervals can be calculated for binomial models.}

\item{...}{Passed on}
}
\value{
Model summary (list)
}
\description{
Determine type of risks model fitted and
generate appropriate summary.
}
\details{
If multiple models were fitted (\code{approach = "all"}), then
the first converged model is displayed. Other models can be accessed
via the returned list \code{$all_models}.
}
