% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postestimation.R
\name{tidy.risks}
\alias{tidy.risks}
\title{Tidy model summaries for risks models}
\usage{
\method{tidy}{risks}(
  x,
  conf.int = TRUE,
  conf.level = 0.95,
  bootrepeats = 100,
  exponentiate = FALSE,
  default = TRUE,
  ...
)
}
\arguments{
\item{x}{Model}

\item{conf.int}{Show confidence intervals?}

\item{conf.level}{Confidence level}

\item{bootrepeats}{Bootstrap repeats. >1000 recommended}

\item{exponentiate}{Exponentiate? For log links}

\item{default}{Use normality-based confidence intervals?}

\item{...}{Passed on}
}
\value{
tibble
}
\description{
Obtain a tibble (data frame) with parameters,
coefficients, standard errors, confidence limits,
and p-values. Usual disclaimers about the use of the latter.
A column with the type of model fitted is also being added.
}
\details{
If running \code{tidy(...)} on a model using marginal
standardization via \code{approach = "margstd"},
by default, the number of bootstrap repeats is very low.
It is strongly recommended to increase them to >1000. See examples.

If multiple types of models are fitted, \code{tidy()} can be used
to parameters for all models at once, in one tibble. The last
column of the tibble includes the name of the model. See examples.
}
\examples{
#fit_rr <- estimate_risk(formula = death ~ stage + receptor, data = dat)
#tidy(fit_rr)

# Increase number of bootstrap repeats:
#fit_rr <- estimate_risk(formula = death ~ stage + receptor, data = dat, approach = "margstd")
#tidy(fit_rr, bootrepeats = 1000)

# Multiple types of models fitted:
#' #fit_rr <- estimate_risk(formula = death ~ stage + receptor, data = dat, approach = "all")
#tidy(fit_rr)
}
